Install configfiles using symlinks
This is very anti-Nix, but duplicates my old dotfile setup. From here it
will be possible to incrementally move configuration into Home Manager.

Note that as of today, this requires using nix version 2.18, which can
be installed with `sudo nix-env -i nix-2.18.1`. It may also be necessary
to explicitly restart the nix-daemon service with `sudo systemctl
daemon-reload` and `sudo systemctl restart nix-daemon.service`. This is
needed due to NixOS/nix#9579, which unfortunately is also not fixed in
nix 2.20.5, which is the new version in nixpkgs. It looks like
NixOS/nix#9723 will fix this and will hopefully land before nix 2.21.