nvim: working treesitter
Currently getting nvim-treesitter via lazy and the parsers via nix. This
runs the risk of having them be out of sync and not working correctly,
as the nvim-treesitter readme notes that it is only guaranteed to work
with the versions of parsers in its lockfile. However, it also gives me
parsers which are certain to load in my nix-built nvim and gives me the
latest queries from nvim-treesitter, so I'm running with it until it
breaks.