nixos: merge nixos config into dotfiles
This adds in the nixos config for snail to my dotfiles. It also moves to
using nixpkgs unstable for snail and for some packages shared between
systems. Notably, it removes logseq and okular.

logseq is no longer in nixpkgs due to being stuck on Electron 27. It is
also currently undergoing a massive rework to move from using markdown
files for storage to using a database. I am not interested in a
proprietary (even if open source) database storage mechanism for my
notes, and I haven't been using it much lately, so this is the time to
cut the cord.

okular is out because there's two versions and I don't know which one is
better. It's been a long time since I last used it.

The final change is to nerd-fonts, which have been reorganized. I've
changed everything to use the unstable nixpkgs version since it should
not matter as they're just fonts.