gaming: use appimage-run and labwc for Path of Exile apps
appimage-run is for Awakened PoE Trade. labwc is for running Path of
Exile and labwc overlayed as its XWayland implementation makes that
easy. I could probably also use a window manager inside raw XWayland,
but this was working faster.

Also, removing Path of Building in favor of the flatpak. I could
probably continue with the nix version and change the wrapper to remove
the WAYLAND_DISPLAY env var as it currently doesn't work correctly in
Wayland. This is probably related to the same NVidia driver bug that's
causing bambu-studio to fail to render in Wayland.