nix: flake update + kernel upgrade on snail
6.13 is no longer supported so this goes to 6.15, which is now supported
by ZFS. New systems package is a dependency of flake-utils, which is a
dependency of nixGL used on non-NixOS systems.

Also dropping the linux-manual package from snail since it depends on
thte latest kernel 6.16 which has a build failure at the moment. How did
this get merged into nixpkgs 25.05 when the kernel entirely fails to
build?