nix: flake update, snail: change to LTS kernel
nvidia-open in nixos 25.05 does not support kernel 6.16, and 6.15 is now
EOL and thus not available. So this goes to 6.12 LTS instead. I'm not
sure whether this supports all of my hardware, in particular monitoring
hardware and 9950X3D scheduling. I haven't benchmarked anything
before/after - this might be worth doing at some point.

With NixOS 25.11, I should be able to use a new kernel again. That would
be a good time to run some benchmarks.