niri: changes to stacked window bindings
This adds H/S bindings for consume-or-expel and also changes all the
bindings to move between monitors and workspaces to move a window, not a
column. I don't group columns by task so I should ralely want this. It's
still available on Mod+Ctrl+Scrollwheel if I really want it, though I
don't use this binding today.