nvim: add norg tree-sitter and neorg
The norg tree-sitter isn't in nvim-treesitter due to having the scanner
written with C++, which was deprecated a while back and is now fully
broken. The neorg dev community hopes to write a new one some day, but
probably only for norg 2.0, which is not yet designed, so it depends on
someone from that community putting much more time into norg
specifically.